Grade 관련 중복 코드 수정
- 정수 형태로 관리되던 Grade를 enum으로 교체
- getGradeImage() 종복 코드를 관련 포로토콜로 처리
